+++ bitcointx/core/secp256k1.py
@@ -27,6 +27,7 @@ import ctypes
 import ctypes.util
 from types import FunctionType
 from typing import Dict, Union, Any, Optional, cast
+from dataclasses import dataclass
 
 import bitcointx.util
 
@@ -66,6 +67,10 @@ _secp256k1_error_storage = Secp256k1LastErrorContextVar(last_error=None)
 _ctypes_functype = getattr(ctypes, 'WINFUNCTYPE', getattr(ctypes, 'CFUNCTYPE'))
 
 
+class secp256k1_context_type:
+    """dummy type for typecheck purposes"""
+
+
 @_ctypes_functype(ctypes.c_void_p, ctypes.c_char_p, ctypes.c_void_p)
 def _secp256k1_illegal_callback_fn(error_str, _data):  # type: ignore
     _secp256k1_error_storage.last_error = {'code': -2, 'type': 'illegal_argument', 'message': str(error_str)}
@@ -88,141 +93,178 @@ def _check_ressecp256k1_void_p(val: int, _func: FunctionType,
     return ctypes.c_void_p(val)
 
 
-secp256k1_has_pubkey_recovery = False
-secp256k1_has_privkey_negate = False
-secp256k1_has_pubkey_negate = False
-secp256k1_has_ecdh = False
-secp256k1_has_xonly_pubkeys = False
-secp256k1_has_schnorrsig = False
+@dataclass(frozen=True)
+class Secp256k1_Capabilities:
+    has_pubkey_recovery: bool
+    has_privkey_negate: bool
+    has_pubkey_negate: bool
+    has_ecdh: bool
+    has_xonly_pubkeys: bool
+    has_schnorrsig: bool
 
 
-def _add_function_definitions(_secp256k1: ctypes.CDLL) -> None:
-    global secp256k1_has_pubkey_recovery
-    global secp256k1_has_privkey_negate
-    global secp256k1_has_pubkey_negate
-    global secp256k1_has_ecdh
-    global secp256k1_has_xonly_pubkeys
-    global secp256k1_has_schnorrsig
+@dataclass(frozen=True)
+class Secp256k1_Contexts:
+    sign: secp256k1_context_type
+    verify: secp256k1_context_type
 
-    if getattr(_secp256k1, 'secp256k1_ecdsa_sign_recoverable', None):
-        secp256k1_has_pubkey_recovery = True
-        _secp256k1.secp256k1_ecdsa_sign_recoverable.restype = ctypes.c_int
-        _secp256k1.secp256k1_ecdsa_sign_recoverable.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p, ctypes.c_char_p, ctypes.c_void_p, ctypes.c_void_p]
 
-        _secp256k1.secp256k1_ecdsa_recoverable_signature_serialize_compact.restype = ctypes.c_int
-        _secp256k1.secp256k1_ecdsa_recoverable_signature_serialize_compact.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.POINTER(ctypes.c_int), ctypes.c_char_p]
+@dataclass(frozen=True)
+class Secp256k1:
+    lib: ctypes.CDLL
+    ctx: Secp256k1_Contexts
+    cap: Secp256k1_Capabilities
 
-        _secp256k1.secp256k1_ecdsa_signature_serialize_compact.restype = ctypes.c_int
-        _secp256k1.secp256k1_ecdsa_signature_serialize_compact.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p]
 
-        _secp256k1.secp256k1_ecdsa_recover.restype = ctypes.c_int
-        _secp256k1.secp256k1_ecdsa_recover.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p, ctypes.c_char_p]
+_secp256k1: Optional[Secp256k1] = None
 
-        _secp256k1.secp256k1_ecdsa_recoverable_signature_parse_compact.restype = ctypes.c_int
-        _secp256k1.secp256k1_ecdsa_recoverable_signature_parse_compact.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p, ctypes.c_int]
 
-    _secp256k1.secp256k1_context_create.restype = ctypes.c_void_p
-    _secp256k1.secp256k1_context_create.errcheck = _check_ressecp256k1_void_p  # type: ignore
-    _secp256k1.secp256k1_context_create.argtypes = [ctypes.c_uint]
+def get_secp256k1() -> Secp256k1:
+    """Will create and initialize an instance of Secp256k1 class, and store
+    it as attribute of the module this function resides in. If this attribute
+    is already initialized, no new instance will be created, and an existing
+    instance will be returned"""
 
-    _secp256k1.secp256k1_context_randomize.restype = ctypes.c_int
-    _secp256k1.secp256k1_context_randomize.argtypes = [ctypes.c_void_p, ctypes.c_char_p]
+    global _secp256k1
 
-    _secp256k1.secp256k1_context_set_illegal_callback.restype = None
-    _secp256k1.secp256k1_context_set_illegal_callback.argtypes = [ctypes.c_void_p, ctypes.c_void_p, ctypes.c_void_p]
+    if _secp256k1 is None:
+        _secp256k1 = secp256k1_load_library(bitcointx.util._secp256k1_library_path)
+        assert _secp256k1 is not None
 
-    _secp256k1.secp256k1_ecdsa_sign.restype = ctypes.c_int
-    _secp256k1.secp256k1_ecdsa_sign.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p, ctypes.c_char_p, ctypes.c_void_p, ctypes.c_void_p]
+    return _secp256k1
 
-    _secp256k1.secp256k1_ecdsa_signature_serialize_der.restype = ctypes.c_int
-    _secp256k1.secp256k1_ecdsa_signature_serialize_der.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.POINTER(ctypes.c_size_t), ctypes.c_char_p]
 
-    _secp256k1.secp256k1_ec_pubkey_create.restype = ctypes.c_int
-    _secp256k1.secp256k1_ec_pubkey_create.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p]
+def _add_function_definitions(lib: ctypes.CDLL) -> Secp256k1_Capabilities:
+    has_pubkey_recovery = False
+    has_privkey_negate = False
+    has_pubkey_negate = False
+    has_ecdh = False
+    has_xonly_pubkeys = False
+    has_schnorrsig = False
 
-    _secp256k1.secp256k1_ec_seckey_verify.restype = ctypes.c_int
-    _secp256k1.secp256k1_ec_seckey_verify.argtypes = [ctypes.c_void_p, ctypes.c_char_p]
+    if getattr(lib, 'secp256k1_ecdsa_sign_recoverable', None):
+        has_pubkey_recovery = True
+        lib.secp256k1_ecdsa_sign_recoverable.restype = ctypes.c_int
+        lib.secp256k1_ecdsa_sign_recoverable.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p, ctypes.c_char_p, ctypes.c_void_p, ctypes.c_void_p]
 
-    _secp256k1.secp256k1_ecdsa_signature_parse_der.restype = ctypes.c_int
-    _secp256k1.secp256k1_ecdsa_signature_parse_der.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p, ctypes.c_size_t]
+        lib.secp256k1_ecdsa_recoverable_signature_serialize_compact.restype = ctypes.c_int
+        lib.secp256k1_ecdsa_recoverable_signature_serialize_compact.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.POINTER(ctypes.c_int), ctypes.c_char_p]
 
-    _secp256k1.secp256k1_ecdsa_signature_parse_compact.restype = ctypes.c_int
-    _secp256k1.secp256k1_ecdsa_signature_parse_compact.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p]
+        lib.secp256k1_ecdsa_signature_serialize_compact.restype = ctypes.c_int
+        lib.secp256k1_ecdsa_signature_serialize_compact.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p]
 
-    _secp256k1.secp256k1_ecdsa_signature_normalize.restype = ctypes.c_int
-    _secp256k1.secp256k1_ecdsa_signature_normalize.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p]
+        lib.secp256k1_ecdsa_recover.restype = ctypes.c_int
+        lib.secp256k1_ecdsa_recover.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p, ctypes.c_char_p]
 
-    _secp256k1.secp256k1_ecdsa_verify.restype = ctypes.c_int
-    _secp256k1.secp256k1_ecdsa_verify.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p, ctypes.c_char_p]
+        lib.secp256k1_ecdsa_recoverable_signature_parse_compact.restype = ctypes.c_int
+        lib.secp256k1_ecdsa_recoverable_signature_parse_compact.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p, ctypes.c_int]
 
-    _secp256k1.secp256k1_ec_pubkey_parse.restype = ctypes.c_int
-    _secp256k1.secp256k1_ec_pubkey_parse.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p, ctypes.c_size_t]
+    lib.secp256k1_context_create.restype = ctypes.c_void_p
+    lib.secp256k1_context_create.errcheck = _check_ressecp256k1_void_p  # type: ignore
+    lib.secp256k1_context_create.argtypes = [ctypes.c_uint]
 
-    _secp256k1.secp256k1_ec_pubkey_tweak_add.restype = ctypes.c_int
-    _secp256k1.secp256k1_ec_pubkey_tweak_add.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p]
+    lib.secp256k1_context_randomize.restype = ctypes.c_int
+    lib.secp256k1_context_randomize.argtypes = [ctypes.c_void_p, ctypes.c_char_p]
 
-    _secp256k1.secp256k1_ec_privkey_tweak_add.restype = ctypes.c_int
-    _secp256k1.secp256k1_ec_privkey_tweak_add.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p]
+    lib.secp256k1_context_set_illegal_callback.restype = None
+    lib.secp256k1_context_set_illegal_callback.argtypes = [ctypes.c_void_p, ctypes.c_void_p, ctypes.c_void_p]
 
-    _secp256k1.secp256k1_ec_pubkey_serialize.restype = ctypes.c_int
-    _secp256k1.secp256k1_ec_pubkey_serialize.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.POINTER(ctypes.c_size_t), ctypes.c_char_p, ctypes.c_uint]
+    lib.secp256k1_ecdsa_sign.restype = ctypes.c_int
+    lib.secp256k1_ecdsa_sign.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p, ctypes.c_char_p, ctypes.c_void_p, ctypes.c_void_p]
 
-    if getattr(_secp256k1, 'secp256k1_ec_pubkey_negate', None):
-        secp256k1_has_pubkey_negate = True
-        _secp256k1.secp256k1_ec_pubkey_negate.restype = ctypes.c_int
-        _secp256k1.secp256k1_ec_pubkey_negate.argtypes = [ctypes.c_void_p, ctypes.c_char_p]
+    lib.secp256k1_ecdsa_signature_serialize_der.restype = ctypes.c_int
+    lib.secp256k1_ecdsa_signature_serialize_der.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.POINTER(ctypes.c_size_t), ctypes.c_char_p]
 
-    if getattr(_secp256k1, 'secp256k1_ec_privkey_negate', None):
-        secp256k1_has_privkey_negate = True
-        _secp256k1.secp256k1_ec_privkey_negate.restype = ctypes.c_int
-        _secp256k1.secp256k1_ec_privkey_negate.argtypes = [ctypes.c_void_p, ctypes.c_char_p]
+    lib.secp256k1_ec_pubkey_create.restype = ctypes.c_int
+    lib.secp256k1_ec_pubkey_create.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p]
 
-    _secp256k1.secp256k1_ec_pubkey_combine.restype = ctypes.c_int
-    _secp256k1.secp256k1_ec_pubkey_combine.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.POINTER(ctypes.c_char_p), ctypes.c_int]
+    lib.secp256k1_ec_seckey_verify.restype = ctypes.c_int
+    lib.secp256k1_ec_seckey_verify.argtypes = [ctypes.c_void_p, ctypes.c_char_p]
 
-    if getattr(_secp256k1, 'secp256k1_ecdh', None):
-        secp256k1_has_ecdh = True
-        _secp256k1.secp256k1_ecdh.restype = ctypes.c_int
-        _secp256k1.secp256k1_ecdh.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p, ctypes.c_char_p, ctypes.c_void_p, ctypes.c_void_p]
+    lib.secp256k1_ecdsa_signature_parse_der.restype = ctypes.c_int
+    lib.secp256k1_ecdsa_signature_parse_der.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p, ctypes.c_size_t]
 
-    if getattr(_secp256k1, 'secp256k1_xonly_pubkey_parse', None):
-        secp256k1_has_xonly_pubkeys = True
-        _secp256k1.secp256k1_xonly_pubkey_parse.restype = ctypes.c_int
-        _secp256k1.secp256k1_xonly_pubkey_parse.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p]
-        _secp256k1.secp256k1_xonly_pubkey_tweak_add_check.restype = ctypes.c_int
-        _secp256k1.secp256k1_xonly_pubkey_tweak_add_check.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_int, ctypes.c_char_p, ctypes.c_char_p]
-        _secp256k1.secp256k1_xonly_pubkey_tweak_add.restype = ctypes.c_int
-        _secp256k1.secp256k1_xonly_pubkey_tweak_add.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p, ctypes.c_char_p]
-        _secp256k1.secp256k1_xonly_pubkey_from_pubkey.restype = ctypes.c_int
-        _secp256k1.secp256k1_xonly_pubkey_from_pubkey.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.POINTER(ctypes.c_int), ctypes.c_char_p]
-        _secp256k1.secp256k1_xonly_pubkey_serialize.restype = ctypes.c_int
-        _secp256k1.secp256k1_xonly_pubkey_serialize.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p]
-        _secp256k1.secp256k1_keypair_create.restype = ctypes.c_int
-        _secp256k1.secp256k1_keypair_create.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p]
-        _secp256k1.secp256k1_keypair_xonly_pub.restype = ctypes.c_int
-        _secp256k1.secp256k1_keypair_xonly_pub.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.POINTER(ctypes.c_int), ctypes.c_char_p]
-        _secp256k1.secp256k1_keypair_xonly_tweak_add.restype = ctypes.c_int
-        _secp256k1.secp256k1_keypair_xonly_tweak_add.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p]
+    lib.secp256k1_ecdsa_signature_parse_compact.restype = ctypes.c_int
+    lib.secp256k1_ecdsa_signature_parse_compact.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p]
+
+    lib.secp256k1_ecdsa_signature_normalize.restype = ctypes.c_int
+    lib.secp256k1_ecdsa_signature_normalize.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p]
+
+    lib.secp256k1_ecdsa_verify.restype = ctypes.c_int
+    lib.secp256k1_ecdsa_verify.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p, ctypes.c_char_p]
+
+    lib.secp256k1_ec_pubkey_parse.restype = ctypes.c_int
+    lib.secp256k1_ec_pubkey_parse.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p, ctypes.c_size_t]
+
+    lib.secp256k1_ec_pubkey_tweak_add.restype = ctypes.c_int
+    lib.secp256k1_ec_pubkey_tweak_add.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p]
+
+    lib.secp256k1_ec_privkey_tweak_add.restype = ctypes.c_int
+    lib.secp256k1_ec_privkey_tweak_add.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p]
+
+    lib.secp256k1_ec_pubkey_serialize.restype = ctypes.c_int
+    lib.secp256k1_ec_pubkey_serialize.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.POINTER(ctypes.c_size_t), ctypes.c_char_p, ctypes.c_uint]
+
+    if getattr(lib, 'secp256k1_ec_pubkey_negate', None):
+        has_pubkey_negate = True
+        lib.secp256k1_ec_pubkey_negate.restype = ctypes.c_int
+        lib.secp256k1_ec_pubkey_negate.argtypes = [ctypes.c_void_p, ctypes.c_char_p]
+
+    if getattr(lib, 'secp256k1_ec_privkey_negate', None):
+        has_privkey_negate = True
+        lib.secp256k1_ec_privkey_negate.restype = ctypes.c_int
+        lib.secp256k1_ec_privkey_negate.argtypes = [ctypes.c_void_p, ctypes.c_char_p]
+
+    lib.secp256k1_ec_pubkey_combine.restype = ctypes.c_int
+    lib.secp256k1_ec_pubkey_combine.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.POINTER(ctypes.c_char_p), ctypes.c_int]
+
+    if getattr(lib, 'secp256k1_ecdh', None):
+        has_ecdh = True
+        lib.secp256k1_ecdh.restype = ctypes.c_int
+        lib.secp256k1_ecdh.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p, ctypes.c_char_p, ctypes.c_void_p, ctypes.c_void_p]
+
+    if getattr(lib, 'secp256k1_xonly_pubkey_parse', None):
+        has_xonly_pubkeys = True
+        lib.secp256k1_xonly_pubkey_parse.restype = ctypes.c_int
+        lib.secp256k1_xonly_pubkey_parse.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p]
+        lib.secp256k1_xonly_pubkey_tweak_add_check.restype = ctypes.c_int
+        lib.secp256k1_xonly_pubkey_tweak_add_check.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_int, ctypes.c_char_p, ctypes.c_char_p]
+        lib.secp256k1_xonly_pubkey_tweak_add.restype = ctypes.c_int
+        lib.secp256k1_xonly_pubkey_tweak_add.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p, ctypes.c_char_p]
+        lib.secp256k1_xonly_pubkey_from_pubkey.restype = ctypes.c_int
+        lib.secp256k1_xonly_pubkey_from_pubkey.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.POINTER(ctypes.c_int), ctypes.c_char_p]
+        lib.secp256k1_xonly_pubkey_serialize.restype = ctypes.c_int
+        lib.secp256k1_xonly_pubkey_serialize.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p]
+        lib.secp256k1_keypair_create.restype = ctypes.c_int
+        lib.secp256k1_keypair_create.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p]
+        lib.secp256k1_keypair_xonly_pub.restype = ctypes.c_int
+        lib.secp256k1_keypair_xonly_pub.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.POINTER(ctypes.c_int), ctypes.c_char_p]
+        lib.secp256k1_keypair_xonly_tweak_add.restype = ctypes.c_int
+        lib.secp256k1_keypair_xonly_tweak_add.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p]
 
     # Note that we check specifically for secp256k1_schnorrsig_sign_custom
     # to avoid incompatibility with earlier version of libsecp256k1.
     # Before secp256k1_schnorrsig_sign_custom was itroduced,
     # secp256k1_schnorrsig_sign had different signature, and using it
     # with this signature will result in segfault.
-    if getattr(_secp256k1, 'secp256k1_schnorrsig_sign_custom', None):
-        secp256k1_has_schnorrsig = True
-        _secp256k1.secp256k1_schnorrsig_verify.restype = ctypes.c_int
-        _secp256k1.secp256k1_schnorrsig_verify.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p, ctypes.c_int, ctypes.c_char_p]
-        _secp256k1.secp256k1_schnorrsig_sign.restype = ctypes.c_int
-        _secp256k1.secp256k1_schnorrsig_sign.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p, ctypes.c_char_p, ctypes.c_char_p]
+    if getattr(lib, 'secp256k1_schnorrsig_sign_custom', None):
+        has_schnorrsig = True
+        lib.secp256k1_schnorrsig_verify.restype = ctypes.c_int
+        lib.secp256k1_schnorrsig_verify.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p, ctypes.c_int, ctypes.c_char_p]
+        lib.secp256k1_schnorrsig_sign.restype = ctypes.c_int
+        lib.secp256k1_schnorrsig_sign.argtypes = [ctypes.c_void_p, ctypes.c_char_p, ctypes.c_char_p, ctypes.c_char_p, ctypes.c_char_p]
+
+    return Secp256k1_Capabilities(
+        has_pubkey_recovery=has_pubkey_recovery,
+        has_privkey_negate=has_privkey_negate,
+        has_pubkey_negate=has_pubkey_negate,
+        has_ecdh=has_ecdh,
+        has_xonly_pubkeys=has_xonly_pubkeys,
+        has_schnorrsig=has_schnorrsig)
 
 
-class _secp256k1_context:
-    """dummy type for typecheck purposes"""
-
-
-def secp256k1_create_and_init_context(_secp256k1: ctypes.CDLL, flags: int
-                                      ) -> _secp256k1_context:
+def secp256k1_create_and_init_context(lib: ctypes.CDLL, flags: int
+                                      ) -> secp256k1_context_type:
     if flags not in (SECP256K1_CONTEXT_SIGN, SECP256K1_CONTEXT_VERIFY,
                      (SECP256K1_CONTEXT_SIGN | SECP256K1_CONTEXT_VERIFY)):
         raise ValueError(
@@ -230,18 +272,18 @@ def secp256k1_create_and_init_context(_secp256k1: ctypes.CDLL, flags: int
             'Must be either SECP256K1_CONTEXT_SIGN, SECP256K1_CONTEXT_VERIFY, '
             'or a combination of these two')
 
-    ctx = _secp256k1.secp256k1_context_create(flags)
+    ctx = lib.secp256k1_context_create(flags)
     if ctx is None:
         raise RuntimeError('secp256k1_context_create() returned None')
 
-    _secp256k1.secp256k1_context_set_illegal_callback(ctx, _secp256k1_illegal_callback_fn, 0)
+    lib.secp256k1_context_set_illegal_callback(ctx, _secp256k1_illegal_callback_fn, 0)
 
     seed = os.urandom(32)
     # secp256k1 commit 6198375218b8132f016b701ef049fb295ca28c95 comment
     # says that "non-signing contexts may use randomization in the future"
     # so we always call randomize, but check for success only for
     # signing context, because older lib versions return 0 for non-signing ctx.
-    res = _secp256k1.secp256k1_context_randomize(ctx, seed)
+    res = lib.secp256k1_context_randomize(ctx, seed)
     if res != 1:
         assert res == 0
         if (flags & SECP256K1_CONTEXT_SIGN) == SECP256K1_CONTEXT_SIGN:
@@ -249,12 +291,14 @@ def secp256k1_create_and_init_context(_secp256k1: ctypes.CDLL, flags: int
         elif flags != SECP256K1_CONTEXT_VERIFY:
             raise AssertionError('unexpected value for flags')
 
-    return cast(_secp256k1_context, ctx)
+    return cast(secp256k1_context_type, ctx)
 
 
-def load_secp256k1_library(path: Optional[str] = None) -> ctypes.CDLL:
+def secp256k1_load_library(path: Optional[str] = None) -> Secp256k1:
     """load libsecp256k1 via ctypes, add default function definitions
-    to the library handle, and return this handle.
+    to the library handle, create 'sign' and 'verify' contexts,
+    and return Secp256k1 class that give access to the handle, contexts,
+    and library capabilities.
 
     Callers of this function must assume responsibility for correct usage
     of the underlying C library.
@@ -280,24 +324,16 @@ def load_secp256k1_library(path: Optional[str] = None) -> ctypes.CDLL:
     except Exception as e:
         raise ImportError('Cannot load secp256k1 library: {}'.format(e))
 
-    _add_function_definitions(handle)
+    cap = _add_function_definitions(handle)
+    ctx = Secp256k1_Contexts(
+        sign=secp256k1_create_and_init_context(handle, SECP256K1_CONTEXT_SIGN),
+        verify=secp256k1_create_and_init_context(handle, SECP256K1_CONTEXT_VERIFY))
 
-    return handle
-
-
-# the handle is not exported purposefully: ctypes interface is low-level,
-# you are basically calling the C functions directly.
-# Anyone that uses it directly should know that they are doing.
-_secp256k1 = load_secp256k1_library(bitcointx.util._secp256k1_library_path)
-
-secp256k1_context_sign = secp256k1_create_and_init_context(_secp256k1, SECP256K1_CONTEXT_SIGN)
-secp256k1_context_verify = secp256k1_create_and_init_context(_secp256k1, SECP256K1_CONTEXT_VERIFY)
+    return Secp256k1(lib=handle, ctx=ctx, cap=cap)
 
 
 __all__ = (
-    'load_secp256k1_library',
-    'secp256k1_context_sign',
-    'secp256k1_context_verify',
+    'secp256k1_load_library',
     'SIGNATURE_SIZE',
     'COMPACT_SIGNATURE_SIZE',
     'PUBLIC_KEY_SIZE',
@@ -306,10 +342,5 @@ __all__ = (
     'SECP256K1_EC_UNCOMPRESSED',
     'SECP256K1_CONTEXT_SIGN',
     'SECP256K1_CONTEXT_VERIFY',
-    'secp256k1_has_pubkey_recovery',
-    'secp256k1_has_privkey_negate',
-    'secp256k1_has_pubkey_negate',
-    'secp256k1_has_ecdh',
-    'secp256k1_has_xonly_pubkeys',
-    'secp256k1_has_schnorrsig',
+    'get_secp256k1',
 )